decoder: post=1 mult as unsigned — closes impulseneg (mult==2^31 sign flip)
OFR_PostProcessor::mult_L/mult_R were declared `int`. mult is read via
read_uniform_split(bit_depth), which for bit_depth=32 can legitimately return values up to
2^32-1; it's a semantically-positive scale factor. Assigning e.g. 2147483648 (2^31, exactly
the mult needed to expand a reduced {-1,0} coded range back to {INT32_MIN,0}, as in a
32-bit "silence + occasional full-negative-scale impulse" signal) into a signed `int`
reinterprets the bit pattern as -2147483648, flipping the sign of every subsequent
`value*mult_L` in OFR_PostProcessor::decode.

Changed the fields to uint32_t (so `(int64_t)value * mult_L` correctly promotes to a
positive int64_t). The one division site that uses mult_L (the post=1 coded-range
reduction in read()) now explicitly casts both operands to int64_t first, since dividing a
signed numerator by an unsigned type would otherwise convert the numerator to unsigned via
the usual arithmetic conversions and break for negative numerators.

Closes the last conformance-suite gap from this session: impulseneg 32-bit, mono+stereo,
all presets. Verified no regression: sine_mono 4c820520, test_stereo 63c303b4, corpus2,
rampslow/noise (this session's earlier 32-bit fixes), and square 8-bit (the original
mult!=1 fix's validation case, mult small and positive, unaffected by the type change).

+++ src/optimfrog_decoder.cpp
@@ -206,12 +206,15 @@ uInt32_t OFR_DecoderEngine::read(void* dest, uInt32_t count) {
                 int mnL = pp->min_val_L, mxL = pp->max_val_L;
                 int mnR = pp->min_val_R, mxR = pp->max_val_R;
                 if (post_type == 1) {
-                    mnL = (pp->min_val_L - pp->offset_L) / pp->mult_L;
-                    mxL = (pp->max_val_L - pp->offset_L) / pp->mult_L;
+                    // mult_L/R is an unsigned scale (may be exactly 2^31); do the division in
+                    // int64_t so a negative numerator stays signed instead of converting to
+                    // unsigned via the usual arithmetic conversions.
+                    mnL = (int32_t)((int64_t)(pp->min_val_L - pp->offset_L) / (int64_t)pp->mult_L);
+                    mxL = (int32_t)((int64_t)(pp->max_val_L - pp->offset_L) / (int64_t)pp->mult_L);
                     if (mnL > mxL) std::swap(mnL, mxL);
                     if (this->channels == 2) {
-                        mnR = (pp->min_val_R - pp->offset_R) / pp->mult_R;
-                        mxR = (pp->max_val_R - pp->offset_R) / pp->mult_R;
+                        mnR = (int32_t)((int64_t)(pp->min_val_R - pp->offset_R) / (int64_t)pp->mult_R);
+                        mxR = (int32_t)((int64_t)(pp->max_val_R - pp->offset_R) / (int64_t)pp->mult_R);
                         if (mnR > mxR) std::swap(mnR, mxR);
                     }
                 }
